Rewrote test buffer functionality.
The new macro `ert-with-test-buffer' replaces both `with-test-buffer'
and `ert-with-temp-buffer-include-file'.  If you need input from
files, combine `ert-with-temp-buffer' with `insert-file-contents'.
The macro `with-test-buffer' was ad-hoc syntactic sugar around
`insert', `goto-char', and `mark', and was not useful enough.

Since test buffers are now linked from the results buffer, there no
longer is a way to browse test buffers.

Also gone is `ert-test-buffer-substitute'.  It had too much overlap
with `ert-filter-string' and was untested and unused.