Cleaned up `ert-simulate-command' somewhat.
Removed the code that was commented out.

Dropped `ert-simulate-command-post-hook' since it's probably not
useful.  Its only use (in nxhtmltest-suites) was rather questionable
since the test code had a permanent side-effect on the value of the hook.

Dropped `ert-simulate-command-delay' since its implementation
(renaming the current buffer to display a temporary message) seemed
too brittle.

Split out `ert-run-idle-timers' as a separate function since what it
does is orthogonal.

If someone wants a function with all the above features, they can
easily write one as a private utility for their own tests; it can call
`ert-simulate-command' and (conditionally) `ert-run-idle-timers', do
whatever would go in the post-hook, display a message, and delay.
Until a convincing implementation of such a combined utility function
emerges, it is better for ERT to provide only the building blocks.